+++ src/main/java/com/example/tae/controller/ProductController.java
@@ -1,0 +1,8 @@
+package com.example.tae.controller;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+
+public class ProductController {
+}
