+++ src/main/resources/static/css/total.css
@@ -17,9 +17,10 @@
 }
 
 .innerGround {
-    margin-top: 40px;
+    margin-top: 30px;
+    margin-right :30px;
     background-color: #ffffff;
-    width: 1470px;
+    width: 1370px;
     height: 1030px;
     border-radius: 100px;
     float: right;
