+++ src/main/resources/static/js/total.js
@@ -414,3 +414,6 @@ document.write(`<script src="/js/ReceivingProcessing.js"></script>`)
 document.write(`<script src="/js/TradingStatement.js"></script>`)
 document.write(`<script src="/js/order.js"></script>`)
 document.write(`<script src="/js/StatusManagementReport.js"></script>`)
+document.write(`<script src="/js/ProductInformationRegistration.js"></script>`)
+document.write(`<script src="/js/ProcurementPlanRegistration.js"></script>`)
+document.write(`<script src="/js/ContractRegistrationModal.js"></script>`)
